requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-SocketIO
Flask-Migrate
Werkzeug
PyJWT
python-dotenv
eventlet
gunicorn